added file "branched_cover.cc" which implements branched covers, and added interface to input_routine.cc.
A branched cover is specified by choosing the branch points, and for each, a
permutation (ie the monodromy around the point). To actually compute the
branched cover, we first need to have a (geometric) packing; we build the
branched cover by "slitting" the plane along straight arcs from the branch
points to (0.1,0.1) and then changing sheets whenever an edge crosses one of
these arcs. Since this computation is done geometrically, we need accurate
data.

To do: implement branched covering "properly" (i.e. directly from the
topology of the graph). Since this step is probably not a bottleneck, there
is no harm in doing it properly (albeit slowly).