made several changes;
(cosmetic change): cast vector sizes to ints as indices in for loops, to prevent comparison of signed and unsigned compile warnings.

(more substantial change): changed the data type of a center_list to a vector of points (instead of two vectors of doubles), and changed the function determine_centers so that it outputs a center_list, and so that the center_list is ordered in vertex order, instead of layout order (note: the center_list is calculated in layout order first, which makes sense, but then reordered to vertex order). As a consequence, several changes were made to other files which assumed center_lists were in layout order.