added function to find vertex closest to a given point (useful for constructing branched covers). To do: rewrite branched_cover.cc, especially the functions new_index and compute_sheet; a plan for how to do this is summarized in the preamble of branched_cover.cc
1. the function new_index should be replaced by an approximation
			new_index(i,s) = s*P.v.size() + i
which is then corrected at the end after all adjacency data is computed.

2. the function compute_sheet tests *every* vertex. instead the new packing
should be written as the trivial branched covering (i.e. (i,s) adjacent to
(j,s) iff i adjacent to j) and the adjusted *only* on indices i which are
on a branch cut, or next to one.